requirements.txt
Django==1.7.3
djangorestframework==3.0.3
wsgiref==0.1.2
